feat: add TopRatedMediaUiState to HomeUiState and integrate top-rated movies use case in HomeViewModel
