+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/home/HomeViewModel.kt
@@ -1,28 +1,25 @@
 package com.berlin.aflami.viewmodel.home
 
-import androidx.lifecycle.viewModelScope
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import com.berlin.aflami.viewmodel.base.BaseViewModel
-import com.berlin.aflami.viewmodel.home.uistate.HomeUiState
 import com.berlin.aflami.viewmodel.base.ErrorUiState
 import com.berlin.aflami.viewmodel.mapper.toUIState
 import com.berlin.aflami.viewmodel.mapper.toUIStateMedia
-import kotlinx.coroutines.Dispatchers
 import com.berlin.aflami.viewmodel.search.GenreUiState
 import com.berlin.aflami.viewmodel.shareduistate.MediaUiState
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import usecase.GetTopRatedMoviesUseCase
-import usecase.GetTopRatedSeriesUseCase
-import kotlinx.coroutines.launch
 import usecase.GetMovieGenresUseCase
 import usecase.GetPopularMoviesUseCase
 import usecase.GetPopularTVShowsUseCase
+import usecase.GetTopRatedMoviesUseCase
+import usecase.GetTopRatedSeriesUseCase
 import usecase.GetUpComingMoviesUseCase
 import usecase.home.GetContinueWatchingMovieUseCase
 import usecase.home.GetContinueWatchingTVShowUseCase
@@ -37,10 +34,6 @@ class HomeViewModel(
     private val getTopRatedSeriesUseCase: GetTopRatedSeriesUseCase,
     private val getTopRatedMoviesUseCase: GetTopRatedMoviesUseCase,
 ) : BaseViewModel<HomeUiState, HomeScreenEffect>(HomeUiState()), HomeInteractionListener {
-    private val getWatchedTVShowUseCase: GetContinueWatchingTVShowUseCase
-) : BaseViewModel<HomeUiState, HomeScreenEffect>(
-    HomeUiState()
-), HomeInteractionListener {
 
     private val _movies = MutableStateFlow<List<MediaUiState>>(emptyList())
     private val _tvShows = MutableStateFlow<List<MediaUiState>>(emptyList())
